+++ iamlistening/handler/_client.py
@@ -27,6 +27,8 @@ class ChatClient:
         Initialize the chat client.
         """
         get = kwargs.get
+        self.__dict__.update(kwargs)
+
         self.name = get("name", None)
         self.enabled = kwargs.get("enabled", None)
 
@@ -100,7 +102,7 @@ class ChatClient:
 
         if self.is_connected:
             self.latest_message = message_content
-            #logger.debug("Frasier👂 on {}: {}", self.platform, message_content)
+            # logger.debug("Frasier👂 on {}: {}", self.platform, message_content)
             await self.handle_iteration_limit()
 
     async def handle_iteration_limit(self):
