+++ iamlistening/handler/matrix.py
@@ -5,7 +5,7 @@ Matrix ⚫
 import simplematrixbotlib as botlib
 from loguru import logger
 
-from .client import ChatClient
+from ._client import ChatClient
 
 
 class MatrixHandler(ChatClient):
