+++ iamlistening/handler/revolt.py
@@ -6,7 +6,7 @@ import aiohttp
 import revolt
 from loguru import logger
 
-from .client import ChatClient
+from ._client import ChatClient
 
 
 class RevoltHandler(ChatClient):
